move deprecated test to tests/deprecated disable Wdeprecated-declarations for deprecated test
