+++ tests/deprecated/test_modifier_coord_scale_old.cpp
@@ -19,7 +19,7 @@
 #include <sys/malloc.h>
 #endif
 
-#include "common_code.hpp"
+#include "../common_code.hpp"
 
 typedef struct
 {
