+++ tests/deprecated/test_modifier_coord_tiny_image_old.cpp
