+++ tests/deprecated/test_modifier_subpix_old.cpp
@@ -19,7 +19,7 @@
 #include <sys/malloc.h>
 #endif
 
-#include "common_code.hpp"
+#include "../common_code.hpp"
 
 typedef struct
 {
